requirements.txt
playwright
pandas
gspread
oauth2client
python-dotenv
schedule
requests
